action_recorder/requirements.txt
pyautogui>=0.9.54
pynput>=1.8.1
Pillow>=11.0.0
numpy>=1.26.4
